requirements.txt
azure-storage-blob
pip install azure-cosmos
python-dotenv
notebook
jupytext
pandas
numpy
requests
tqdm